% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\docType{data}
\name{CD}
\alias{CD}
\title{Census Division data set with geometry}
\format{
A data frame with 293 rows and 7 variables:
\describe{
\item{CDNAME}{Census Division names}
\item{PRUID}{ID column for each province}
\item{rmapshaperid}{id for geometry}
\item{population_2021}{Population from 2021 Census}
\item{PT}{province or territory}
\item{PRNAME}{Province name}
\item{geometry}{map geometry}
...
}
}
\usage{
CD
}
\description{
A data set containing the Canadian Census Division geometry shapes
}
\keyword{datasets}
