% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{coord_transform}
\alias{coord_transform}
\title{mapping points coordinates}
\usage{
coord_transform(data, long, lat)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{long}{the longitude variable name}

\item{lat}{the latitude variable name}
}
\value{
Your coordinates transformed.
}
\description{
This function converts your coordinates to match the mapping to overlay in ggplot.
}
