% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{crs_coord}
\alias{crs_coord}
\title{ggplot map coordinate system}
\usage{
crs_coord()
}
\value{
a coordinate system for ggplot2
}
\description{
Used to visualise simple feature (sf) objects. Required to plot geom objects.
}
