% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\docType{data}
\name{FSA}
\alias{FSA}
\title{Canadian FSA Boundary Data}
\format{
A data frame with 1614 rows and 6 variables:
\describe{
\item{CFSAUID}{ID column for each FSA code}
\item{PRNAME}{Province name}
\item{PRUID}{ID column for each province}
\item{PT}{province or territory}
\item{rmapshaperid}{id for geometry}
\item{geometry}{map geometry}
...
}
}
\usage{
FSA
}
\description{
Provides geospatial information for all Canadian Forward sortation areas (FSA).
}
\keyword{datasets}
