% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_cd}
\alias{geom_cd}
\title{Mapping Census Division data}
\usage{
geom_cd(data = CD, fill = "PRNAME", colour = "white", size = 0.2)
}
\arguments{
\item{data}{a data set with geometry variable}

\item{fill}{the color fill variable}

\item{colour}{outline color, default is NA}

\item{size}{size of outline}
}
\value{
FSA map.
}
\description{
Maps CD data using Statistics Canada CD shape file.
}
