% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_csd}
\alias{geom_csd}
\title{Mapping Census Sub Division data}
\usage{
geom_csd(data = CSD, fill = "PRNAME", colour = "white", size = 0.2)
}
\arguments{
\item{data}{a data set with geometry variable}

\item{fill}{the color fill variable}

\item{colour}{outline color, default is NA}

\item{size}{size of outline}
}
\value{
FSA map.
}
\description{
Maps CSD data using Statistics Canada CSD shape file.
}
