% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_fsa}
\alias{geom_fsa}
\title{Mapping FSA data}
\usage{
geom_fsa(data = FSA, fill = "PRNAME", colour = "white", size = 0.2)
}
\arguments{
\item{data}{a dataset with geometry variable}

\item{fill}{the colour fill variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
FSA map.
}
\description{
Maps FSA data using statscan FSA shapefile.
}
