% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_hr}
\alias{geom_hr}
\title{Mapping HR data}
\usage{
geom_hr(data = HR, fill = "PRNAME", colour = "white", size = 0.2)
}
\arguments{
\item{data}{a data set with geometry variable}

\item{fill}{the color fill variable}

\item{colour}{outline color, default is NA}

\item{size}{size of outline}
}
\value{
HR map.
}
\description{
Maps HR data using Statistics Canada HR shape file.
}
