% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_prov}
\alias{geom_prov}
\title{Mapping provincial data}
\usage{
geom_prov(data = PROV, fill = "PT", colour = NA, size = 0.1)
}
\arguments{
\item{data}{a data set with geometry variable}

\item{fill}{the color fill variable}

\item{colour}{outline color, default is NA}

\item{size}{size of outline}
}
\value{
Provincial map.
}
\description{
Maps provincial data using Statistics Canada province shape file.
}
