% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{geom_reg}
\alias{geom_reg}
\title{mapping regional shapefile}
\usage{
geom_reg(data = REG, fill = "region", colour = NA, size = 0.1)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{fill}{the colour fill variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
Provincial map.
}
\description{
This function maps the regional shapefile.
}
