% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\docType{data}
\name{HR}
\alias{HR}
\title{Canadian HR Boundary Data}
\format{
A data frame with 104 rows and 8 variables:
\describe{
\item{HRID}{ID column for each HR code}
\item{HRENGNAME}{Health Region english name}
\item{HRFRNAME}{Health Region french name}
\item{rmapshaperid}{id for geometry}
\item{geometry}{map geometry}
\item{PT}{province or territory}
\item{PRNAME}{Province name}
\item{PRUID}{ID column for each province}
...
}
}
\usage{
HR
}
\description{
Provides geospatial information for all Health Regions areas (HR).
}
\keyword{datasets}
