% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{label_reg}
\alias{label_reg}
\title{adding regional labels}
\usage{
label_reg(data = REG, label = "region", colour = "grey20", size = 4)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{label}{the label variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
Regional map labels.
}
\description{
This function adds labels in the center of each region.
}
