% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\docType{data}
\name{REG}
\alias{REG}
\title{Canadian Regions Boundary Data}
\format{
A data frame with 13 rows and 12 variables:
\describe{
\item{PREABBR}{Province abbreviated English}
\item{PRENAME}{Province name English}
\item{PRFABBR}{Province abbreviated French}
\item{PRFNAME}{Province name French}
\item{PRNAME}{Province name}
\item{PRUID}{ID column for each province}
\item{PT}{province or territory}
\item{region}{national region}
\item{X}{center of shape x axis}
\item{Y}{center of shape y axis}
\item{rmapshaperid}{id for geometry}
\item{geometry}{map geometry}
...
}
}
\usage{
REG
}
\description{
Provides geospatial information for all 4 regions of Canada (Central, Eastern, Western, and Northern).
}
\keyword{datasets}
