% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{scale_fill_map}
\alias{scale_fill_map}
\title{Color Palettes Fill}
\usage{
scale_fill_map(palette, num, na.value = "grey90", rev = FALSE)
}
\arguments{
\item{palette}{color palette name}

\item{num}{number of colors to create}

\item{na.value}{a color value for NA, defaults to light gray}

\item{rev}{condition to reverse the order of the colors}
}
\value{
a ggplot color palette
}
\description{
Create a custom number of colors to use for graphing or mapping based on preset color palettes.
}
