% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{text_prov}
\alias{text_prov}
\title{Province text}
\usage{
text_prov(data = PROV, label = "PT", colour = "grey20", size = 3)
}
\arguments{
\item{data}{a dataset with geometry variable}

\item{label}{the label variable}

\item{colour}{text colour}

\item{size}{text size}
}
\value{
Provincial map labels
}
\description{
Adds text labels in the center of each province.
}
