% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{theme_map}
\alias{theme_map}
\title{Plot theme - map}
\usage{
theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{size for text}

\item{base_family}{font for text}
}
\value{
A blank theme for ggplot
}
\description{
Blank theme to apply for mapping using ggplot2.
}
