% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{coord_transform}
\alias{coord_transform}
\title{Transforming map coordinates}
\usage{
coord_transform(data, long, lat)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{long}{the longitude variable name}

\item{lat}{the latitude variable name}
}
\value{
Your coordinates transformed.
}
\description{
Converts your longitude and latitude coordinates to match the maps to properly overlay in ggplot.
}
