% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{crs_coord}
\alias{crs_coord}
\title{Map coordinate system}
\usage{
crs_coord()
}
\value{
a coordinate system for mapping
}
\description{
Used to visualize simple feature (sf) objects. Required to plot geometry objects.
}
