% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\docType{data}
\name{CSD}
\alias{CSD}
\title{Census Sub Division data set with geometry}
\format{
A data frame with 5161 rows and 9 variables:
\describe{
\item{CDNAME}{Census Division names}
\item{PRUID}{ID column for each province}
\item{CDUID}{Census Division ID}
\item{CMAUID}{CMA ID}
\item{rmapshaperid}{id for geometry}
\item{population_2021}{Population from 2021 Census}
\item{PT}{province or territory}
\item{PRNAME}{Province name}
\item{geometry}{map geometry}
...
}
}
\usage{
CSD
}
\description{
A data set containing the Canadian Census Sub Division geometry shapes
}
\keyword{datasets}
