% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\docType{data}
\name{FSA}
\alias{FSA}
\title{FSA data set with geometry}
\format{
A data frame with 1614 rows and 6 variables:
\describe{
\item{CFSAUID}{ID column for each FSA code}
\item{PRNAME}{Province name}
\item{PRUID}{ID column for each province}
\item{PT}{province or territory}
\item{rmapshaperid}{id for geometry}
\item{geometry}{map geometry}
...
}
}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2016-eng.cfm}
}
\usage{
FSA
}
\description{
A data set containing the Canadian FSA geometry shapes
}
\keyword{datasets}
