% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{geom_fsa}
\alias{geom_fsa}
\title{mapping fsa shapefile}
\usage{
geom_fsa(data = FSA, fill = FSA$PRNAME, colour = "white", size = 0.2)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{fill}{the colour fill variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
FSA map.
}
\description{
This function maps the FSA shapefile.
}
