% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{geom_prov}
\alias{geom_prov}
\title{mapping provincial shapefile}
\usage{
geom_prov(data = PROV, fill = "PT", colour = NA, size = 0.1)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{fill}{the colour fill variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
Provincial map.
}
\description{
This function maps the provincial shapefile.
}
