% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_reg}
\alias{geom_reg}
\title{Mapping regional data}
\usage{
geom_reg(data = REG, fill = "region", colour = NA, size = 0.1)
}
\arguments{
\item{data}{a dataset with geometry variable}

\item{fill}{the colour fill variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
Regional map.
}
\description{
Maps regional data using statscan province shapefile.
}
