% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{geom_reg}
\alias{geom_reg}
\title{Mapping regional data}
\usage{
geom_reg(data = REG, fill = "region", colour = NA, size = 0.1)
}
\arguments{
\item{data}{a data set with geometry variable}

\item{fill}{the color fill variable}

\item{colour}{outline color, default is NA}

\item{size}{size of outline}
}
\value{
Regional map.
}
\description{
Maps regional data using Statistics Canada province shape file.
}
