% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{label_prov}
\alias{label_prov}
\title{Province labels}
\usage{
label_prov(
  data = PROV,
  label = "PT",
  colour = "grey20",
  size = 3,
  outline = NA,
  alpha = 0.7
)
}
\arguments{
\item{data}{a dataset with geometry variable}

\item{label}{the label variable}

\item{colour}{label colour}

\item{size}{label size}
}
\value{
Provincial map labels
}
\description{
Adds text labels in the center of each province with light grey bubble.
}
