% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{label_prov}
\alias{label_prov}
\title{adding labels}
\usage{
label_prov(data = PROV, label = "PT", colour = "grey20", size = 3)
}
\arguments{
\item{data}{a dataset with long and lat coordinates}

\item{label}{the label variable}

\item{colour}{outline colour, default is NA}

\item{size}{size of outline}
}
\value{
Provincial map.
}
\description{
This function adds labels in the center of each province.
}
