% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{label_reg}
\alias{label_reg}
\title{Regional labels}
\usage{
label_reg(data = REG, label = "region", colour = "grey20", size = 4)
}
\arguments{
\item{data}{a data set with geometry variable}

\item{label}{the label variable}

\item{colour}{label color}

\item{size}{label size}
}
\value{
Regional map labels.
}
\description{
Adds text labels in the center of each region with light gray bubble.
}
