% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadianmaps.R
\name{scale_color_map}
\alias{scale_color_map}
\title{Colour Palettes Line Color}
\usage{
scale_color_map(palette, num, na.value = "grey90", rev = FALSE)
}
\arguments{
\item{palette}{colour palette name}

\item{num}{number of colours to create}

\item{na.value}{a colour value for NA, defaults to light grey}
}
\value{
a ggplot colour palette
}
\description{
Create a custom number of colours to use for graphing or mapping based on preset colour palettes.
}
