% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadamaps.R
\name{scale_fill_map}
\alias{scale_fill_map}
\title{Colour Palettes}
\usage{
scale_fill_map(palette, num, na.value = "grey90")
}
\arguments{
\item{palette}{colour palette name}

\item{num}{number of colours to create}

\item{na.val}{a colour value for NA, defaults to light grey}
}
\value{
a ggplot colour scale
}
\description{
This function creates colour palettes.
}
